refactor(auth): relocate auth provider handlers to services
