+++ apps/api/tests/qoder-provider.test.ts
@@ -3,7 +3,7 @@ import { afterEach, test } from "node:test";
 import { deleteProviderDB, getProviderByIdDB, upsertProviderDB } from "@srouter/db";
 import type { ProviderConfig } from "@srouter/types";
 import { AuthLogic } from "../src/logic/auth.logic.js";
-import { qoderAuthHandler } from "../src/logic/auth.providers.js";
+import { qoderAuthHandler } from "../src/services/authHandlers.js";
 
 const createdIds: string[] = [];
 const originalFetch = globalThis.fetch;
