+++ apps/api/tests/tokenrouter-provider.test.ts
@@ -3,7 +3,7 @@ import { afterEach, test } from "node:test";
 import { deleteProviderDB, upsertProviderDB } from "@srouter/db";
 import type { ProviderConfig } from "@srouter/types";
 import { AuthLogic } from "../src/logic/auth.logic.js";
-import { tokenRouterAuthHandler } from "../src/logic/auth.providers.js";
+import { tokenRouterAuthHandler } from "../src/services/authHandlers.js";
 
 const createdIds: string[] = [];
 const originalFetch = globalThis.fetch;
